pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]
asyncio_mode="auto"

[tool.ruff]
line-length = 120

[tool.ruff.per-file-ignores]
"app/db/models/task.py" = ["F821"]
"app/db/models/user.py" = ["F821"]

[tool.poetry]
name = "fastapi-task-manager"
version = "0.1.1"
description = "FastAPI Task Manager application"
authors = ["m4xPatrol"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.21"}
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.28.0"
pydantic = {extras = ["email"], version = "^2.4.2"}


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
